Don't resolve provided deps, don't require a resource folder
The maven resolver can be mutated in a way that it resolves
dependencies with the provided scope if it is needed, however
in most cases it is not.

Furthermore this change vastly improves lookup speeds though
to be honest this is more or less a coverup and I should really
be looking at why resolution takes multiple seconds. My uneducated
guess says that it is the document parsing and iteration that is
causing this but I'd really need to look at why exactly it is
doing that